Extract custom assertions to a separate file\
